pyproject.toml
[project]
name = "dental-agent-simple"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "elevenlabs>=2.16.0",
    "google-api-python-client>=2.184.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "loguru>=0.7.3",
    "openai>=1.74.0,<=1.99.1",
    "pipecat-ai[webrtc,soniox,websocket,elevenlabs,daily,silero,deepgram,cartesia,openai,anthropic,google,aws,runner]>=0.0.85",
    "pipecat-ai-flows>=0.0.21",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=21.0",
]
